tsar: change the way IRQs are handled
Before, we had a driver for the mips32 internal icu. It seems
unnecessary since we always use an external icu (vci_icu, or vci_xicu
for SMP), so discard it. We always unmask the IRQ sources of the mips32.
Also the system icu (vci_icu or vci_xicu) now provide the first level
IRQ handler, which must demux the pending IRQ and call the generic
do_IRQ function.