tsar: rework trap installation
Before we were allocating a page at runtime to install a two line
trampoline, to jump to the exception vector. Now, we just compile
directly the exception vector at a proper offset/alignment so there is
no need to allocate wasted memory at runtime. We also separate the
exception handlers installation from the cpu configuration (setup of the
exception address and user access to certain system registers). It will
be easier for SMP support since the exception handlers installation
needs to be done only once, while the cpu configuration needs to be
performed on every cpu.