tsar: vci_xicu: rework irq number when mapping hw irqs
irqs #0 and #1 are reserved for IPI and private timer (percpu) in the
xicu irq domain. But we shouldn't force using irq number >= 2 in the
device tree. This patch makes it possible to declare hw irq number
starting from 0 in the device tree for other peripherals, and we add 2
ourselves in the irq mapping (this way it's more generic with respect to
the number of private virtual irqs we want to define).