tsar: vci_mtty: only test one bit for VCI_STATUS
With SoCLib, VCI_STATUS is basically a bool: ==0 if no pending
character, !=0 otherwise.
With FPGA, VCI_STATUS is more complex and contains more relevant bits.

So, in both cases, to know if a pending character is present in the
input buffer, test only the LSB of VCI_STATUS.

Weirdly, a consequence of this patch is to make panic messages appear
way faster on FPGA.