tsar: add support for LETI system on SoCLib
The LETI system is very similar to FPGA: almost the same address mapping
but different interrupt lines.

Also, in this system, the bootloader is loaded in 0x0, so we have to
load the kernel further away from the beginning of the memory (i.e. we
load it at 0x8000 == 32KiB).

For now, it only supports 1 CPU.