tsar: add argument to smp ipi calls
Normally, we use the ipi mailbox as a simple flag to raise an interrupt.
The content of the ipi message is given by a regular structure in
memory.

For smp ipi boot (when using the tsar preloader), the ipi mailbox must
receive the address of where the targeted slave processor should jump
to. Prepare for such a behavior.