tsar: smp boot by ipi
Instead of having all the processors (including slaves) boot at the same
time, and only the master processor doing the real job while the other
are waiting on a signature variable, this patch makes it possible to
wake slave processors via ipi only when we need to boot them.

The preloader configures the xicu to enable IPI for slave processor.
When the master processor wants to boot a slave processor, it has to
send an IPI message containing the address the slave has to jump to. In
our case, we assume the boot processor is always processor 0. The
address we give to slave processors is the regular entry point (the same
as the master), but they are quickly redirect to a secondary entry
point (see kernel/head.S).