tsar: introduce a new macro for PTPR shifting
The PTPR (page table pointer register) value has to be 13-bit right
shifted, since the page table is aligned on a 8KiB boundary anyway. So
let us hardcode this shifting value, so we don't have to make the shift
computation each time.

Also, export the macro to assembly (e.g. for use in head.S).