tsar: avoid making the shifting computation of PTPR in enable_paging()
Before, the full physical address of the page table was given to
enable_paging() which was then performing the 13-bit right shift before
writing the PTPR in the MMU.

Now, for 40-bit support and to avoid dealing with giving a 40-bit value
to enable_paging() as a pointer to a page table, we perform the right
shift on the value beforehand. So the value that enable_paging()
receives is already shifted and thus can be contained in a simple 32-bit
word.