tsar: save the current thread_info in a system register
When switching to kernel mode from user mode, we have to find the kernel
stack associated to the thread. This data is deduced from getting hold
of the thread's thread_info structure.

Until now, a pointer on the current thread_info structure (for a given
cpu) was saved in array, indexed by the logical index of the given cpu.
The problem is that when switching to kernel mode, all we can get is the
hwcpuid which we then should transform into the logical cpuid. This
would be a freaking pain, especially with clusterized TSAR, where
hwcpuid can be very sparse (basically, we'd need to provide an
1024-entry array, to provide for the 10 possible bits of the
EBASE[HWCPUID]).

Instead of using such a 'software' array, we now use a cpu system
register. When switching to a new context, and already setting $gp to
point on the current thread_info, we also setup a system register
(tccontext) to store the same value. When switching from user mode to
kernel mode, we just have to read this value to get the proper
thread_info structure.