tsar: remove support for early fixmap
Early fixmap was only used for the early mapping of the tty
and allow earlyprintk to work.

But with the introduction of HIGHMEM/NUMA support, it's actually a bit
of a pain to keep maintaining the early fixmap support (see the x86 code
to see the mess) especially if it's really just for one device.

Since we found a way to access the tty without the need for early fixmap
(ie by temporarily deactivating the MMU and accessing the tty using its
physical address), we can just remove it altogether.