tsar: add CONFIG_HIGHMEM infrastructure
HIGHMEM enables the use of memory that is located above the lowmem
limit, ie the first ~900MiB in kernel space.

HIGHMEM is necessary for the first version of TSAR NUMA support, since
all the memory banks other than the cluster 0 are above the lowmem limit
in the physical address space.

Later, for the second versio of NUMA support, when we'll be able to
stack the memory blocks in the kernel space, it won't be as necessary
but still useful when reaching a certain number of clusters (ie with a
total memory size > 900MiB).