tsar: vci_mtty: add a comment about the virtual irq initialization
When probing the tty, we get a virtual irq number for the irq line
connecting the tty to its associated icu. I couldn't remember how this
number was obtained, and started freaking out.

Turns out the virtual irq number is computed beforehand when populating
the platform, using the proper of_* functions and thus operating the
right irq domain mapping. Phew!