tsar: vci_xicu: adaptation to NUMA
Make it possible to have one XICU/timer per cluster. They are
initialized at boot by the boot cpu, when analysed from the device tree.

---

For now, IPI and PTI are still declared as percpu interrupt even though
one "logical" IPI is actually local to a cluster and not to the whole
system as before. There will thus be nr_nodes * IPI in the system (and
the same of PTI).

The percpu attribute could be misleading. The only advantage that I see
is that we use 4 times less irqdesc than if we used one IRQ per IPI/PTI
per processor. Is it worth it?

Eventually, it could make more sense to have as many IPIs/PTIs as
processors, and just make them non-balanceable.