tsar: new driver for the XICU/IOPIC component
The IOPIC is a kind of interrupt controller, mainly intended for LETI
systems. It is meant to be located in the IO cluster, and forwards
interrupts from devices in the IO cluster to cluster-local XICUs via
regular network messages (instead of using a classic interrupt network).

Here, the IOPIC driver is closely bound to the XICU driver (even though
the IOPIC theoritically is a standalone component only forwarding
interrupts to any physical address - which doesn't necessarily imply
using a XICU as backend) because 1/ that's how our systems work, 2/ it
wouldn't make much sense not to do so.

Interrupts coming from the IOPIC can migrate on any processor of the
system (although at first, they'll all be forwarded to the bootup
processor).

And finally, we apply a linear mapping between IOPIC IRQs and XICU WTIs,
which means that the 32 possible IOPIC IRQs are not available since some
XICU WTIs are already reserved for IPIs (one per processor, at
cluster-level). Assuming there are 4 processors/cluster, then it leaves
28 free WTIs, onto which  we can map the first 28 IOPIC IRQs.