tsar: fix the condition that warns whenever a node is already online
We assume that each node can only contain one memory bank. So when we
associate memory banks to nodes, we warn if ever the corresponding node
is already online. The node #0 is by default online, so we don't have to
check when adding the first memory bank to the node #0.