tsar: add nop operations after writing values in coprocessor 2
In the CEA LETI model of the processor, values are sent to the
coprocessor 2 in the MEM stage of the pipeline. For example when
activating the MMU (ie through a write in CP2), we must delay the next
instructions so they're not performed when the MMU is still off (eg JUMP
target addresses are computed in DEC stage).