tsar: add arch_cpu_idle callback
By default, without a specific arch-specific implementation of idle(),
cpus do polling which destroys the performance of a multi-cluster system
(all cpus try to access memory in cluster 0).

Simple implementation of idle() by using the MIPS32 instruction "wait"
which switches a cpu into low-power mode and make it wait for an
external event to happen (ie an interruption).