tsar: ioc: the blockdevice cannot handle several simultaneous requests
Fix bug in the request function of the IOC driver. Instead of having a
loop on the request queue and sending every requests to the IOC, we can
send only one at a time. The calling Linux function is in charge of
looping through the request queue and calling our function when the
driver is finally available.