tsar: enable irqs in do_notify_resume
IRQs must be enabled in do_notify_resume, otherwise it can trigger a
irqs_disabled() check later on.

However, current_thread_info->flags must be read before irqs are
enabled. We thus read it in entry.S (and put them in a1), so
do_notify_resume() gets it as an argument when being called, before irqs
are enabled.