tsar: mm: introduce lowmem stacking across nodes
By default in 32-bit Linux, only the physical memory starting from 0x0
to ~800MB gets mapped in the lowmem region of the kernel space. Any
physical memory ranges above remains accessible but only through the
highmem (ie costly temporary mappings).

In TSAR, the 40-bit physical address space is split across nodes with
the 8 MSB describing XY coordinates of nodes (the 32 LSB being the local
physical address space inside a node). It means, that in the default
configuration, we're are only able to map the physical memory bank of
node #0 in lowmem and all the memory of the other nodes are stuck in
highmem.

In this patch, we tweak the mapping of the lowmem so that it is not
linear anymore. We can stack up memory of several nodes, which
physical ranges are discontinuous, linearly in the lowmem.

If the sum of all the memory banks is higher than the hardlimit for
lowmem, we have two options:

- reducing the size of each memory bank that we want to map in lowmem.
  The remaining will then be put in highmem. We cannot reduce under a
  certain size though (which we decided would be the maximum between the
  kernel size - the kernel must be contiguously mapped - and MAX_ORDER).

- reducing the number of nodes that we want to map in lowmem. For
  example, we can map only half of all the nodes in lowmem. The rest
  will be in highmem.