tsar: idmap: make the idmap argument to secondary cpus more generic
Before, secondary cpus were directly fetching a global variable defined in
mm/idmap.c to get the idmap page table physical address.

Now, this physical address is passed to secondary cpus as a generic argument, so
that in the future generation of the idmap page table is better separated from
secondary cpu booting.