tsar: add kernel code and rodata replication
In a massively multi-processor system, access to memory is one of the biggest
bottlenecks. And if all the kernel text and data are located in the same memory
bank, then it means all the processors of the system will try to access this
bank at the same time and all the time!

In simulation, we found out that in a 96-processor system, when the last cluster
is booting, then the latency to memory reaches up to 25000 cycles per memory
request.

In this patch, we replicate the code and rodata of the kernel, in order to
unload the memory bank containing the original copy.

First, we determine the number of replication copies there will be and where
they will be located. Since we are willing to allocate a maximum 6.25% of the
lowmem memory for replication (it's easy on the cpu: 100 / 16 = 100 >> 4 =
6.25%), the heuristic is to lower the number of nodes which will contain a copy
until it fits the memory requirement.

Then we build a set of patched page tables, based on swapper_pg_dir, each
targeting a replication copy of the kernel code and rodata.

Finally when secondary cpus boot up, we give them each the page table that
target the closest copy.

Note that these page tables are only relevant during boot, so all the cpus don't
try to access the same memory bank. Then, during normal execution, cpus won't
use them anymore but will reuse userspace processes' page tables.