tsar: mm: extend the kernel memblock reservation
Before, we used to reserve from _stext to _end, thus excluding the assembly boot
code (ie .head.text). But even if the area containing the boot code is not
likely to be allocated, let's mark it as reserved as well, at least by safety.
After all, we still need the boot code to exist when we boot the secondary cpus.