tsar: the DTB is now placed by the bootloader before the kernel image in memory
Before, the DTB could actually be placed anywhere in memory (more accurately
anywhere within the address space of the cluster #0). But:

1/ Linux cannot access the entirety of this address space once it has turned on
virtual memory, because at first only the kernel image is mapped. And to map the
rest, Linux has to have access to the DTB.

2/ Linux could access the DTB while still running in physical mode, and copy it
somewhere more convenient. It is difficult to find a location that is
convenient though, since memory allocation is impossible at this point, as well
as error detection and reporting.

Now, Linux is compiled with a fixed offset of 256 KiB from the start of memory
which should provide enough space for the biggest TSAR DTB. It is up to the
bootloader to copy the DTB before the kernel image, and to report any errors.

Linux only has to initially map from address 0 to the end of the kernel image,
and gains access to the DTB automatically. Then it has to reserve the memory
used for the DTB so it is not freed at the end of init.