tsar: xicu: avoid unnecessary bug reporting when SMP && !MULTI_CLUSTER
During initialization, the XICU driver tests that the XICU hardware component
has enough internal timers, external IRQ line outputs, etc. to accommodate the
number of processors.

When multicluster, the number of processors per cluster is fixed at 4. But in
monocluster SMP setups, NR_CPUS was used which is in most cases too big. We
change that for num_possible_cpus() which gives us an accurate number.